refactor: EvalFactory now defined in task.py
